requirements.txt
python-dotenv~=0.14.0
tweepy~=3.9.0
pandas~=1.1.2
asyncpg~=0.21.0
requests~=2.24.0